Move PipelineAdapter to `pipelines` module
