+++ fedot/core/pipelines/adapters.py
