+++ fedot/core/pipelines/verification.py
@@ -11,7 +11,7 @@ from fedot.core.dag.verification_rules import (
     has_no_self_cycled_nodes,
     has_one_root
 )
-from fedot.core.optimisers.adapters import PipelineAdapter
+from fedot.core.pipelines.adapters import PipelineAdapter
 from fedot.core.optimisers.graph import OptGraph
 from fedot.core.pipelines.verification_rules import (
     has_correct_data_connections,
