+++ test/unit/dag/test_graph_operator.py
@@ -2,7 +2,7 @@ from copy import deepcopy
 
 from fedot.core.dag.linked_graph import LinkedGraph, get_distance_between
 from fedot.core.dag.graph_utils import distance_to_root_level, nodes_from_layer
-from fedot.core.optimisers.adapters import PipelineAdapter
+from fedot.core.pipelines.adapters import PipelineAdapter
 from fedot.core.optimisers.graph import OptNode
 from fedot.core.pipelines.node import PrimaryNode, SecondaryNode
 from fedot.core.pipelines.pipeline import Pipeline
